% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{count_bigrams}
\alias{count_bigrams}
\title{Bigrams dataframe tokenizer}
\usage{
count_bigrams(dataset)
}
\arguments{
\item{dataset}{a dataframe with the lyrics}
}
\value{
A tidytext dataframe
}
\description{
Tokenize lyrics in bigrams and group by the bigrams.
}
\examples{

}
