% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_acquisition.R
\name{get_artist_songs_in_df}
\alias{get_artist_songs_in_df}
\title{Retrieve songs of artist}
\usage{
get_artist_songs_in_df(artist_query)
}
\arguments{
\item{artist_query}{the string containing the artist name}
}
\value{
A dataframe with all the songs of an artist and other infos(like song_id)
}
\description{
Searches and gets all songs of a specified artist
}
\examples{

}
