% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_acquisition.R
\name{get_lyric_from_song}
\alias{get_lyric_from_song}
\title{Retrieve lyrics of a song}
\usage{
get_lyric_from_song(song_name, songs_df)
}
\arguments{
\item{song_name}{A string with the song name}

\item{songs_df}{A dataframe with songs names and songs Genius ids}
}
\value{
A dataframe with all the lyric of the song. Each row of the dataframe is a line
}
\description{
gets lyric given a song and a songs dataframe (that includes Genius song id)
}
\examples{

}
