% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_acquisition.R
\name{get_lyrics_from_songs_df}
\alias{get_lyrics_from_songs_df}
\title{Retrieve lyrics of multiple songs}
\usage{
get_lyrics_from_songs_df(songs_df)
}
\arguments{
\item{songs_df}{A dataframe with songs names and songs Genius ids}
}
\value{
A list of dataframes with all the lyrics of the songs. Each row of the dataframe is a line
}
\description{
gets lyrics from a dataframe of songs
}
\examples{

}
