% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_acquisition.R
\name{get_my_current_play}
\alias{get_my_current_play}
\title{Retrieve currently playing song on Spotify}
\usage{
get_my_current_play()
}
\value{
An array of various information about the currently playing song
}
\description{
gets the currently playing song on Spotify
}
\examples{

}
