% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{make_lda_lyrics}
\alias{make_lda_lyrics}
\title{2-topic LDA for lyrics}
\usage{
make_lda_lyrics(lyrics_tokens_clean)
}
\arguments{
\item{tokens_single_clean}{A tidytext dataframe already made of unigrams and clean from stopwords}
}
\value{
A tidytext dataframe
}
\description{
Makes the Latent Dirichlet allocation for the specified lyric
}
\examples{

}
