% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{sentiment_lyrics_count}
\alias{sentiment_lyrics_count}
\title{Lyrics sentiment joiner}
\usage{
sentiment_lyrics_count(tokens_single_no_stop)
}
\arguments{
\item{tokens_single_no_stop}{A tidytext dataframe already made of unigrams and clean from stopwords}
}
\value{
A tidytext dataframe
}
\description{
Joins to the tidytext unigrams dataframe the value of the sentiment (bing database) and group by words
}
\examples{

}
