% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{tokenize_lyrics}
\alias{tokenize_lyrics}
\title{Lyrics tokenizer}
\usage{
tokenize_lyrics(single_lyric)
}
\arguments{
\item{single_lyric}{a dataframe with the lyrics}
}
\value{
A tidytext dataframe
}
\description{
Tokenize lyrics in unigrams
}
\examples{

}
