% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{tokenize_lyrics_clean_stopwords}
\alias{tokenize_lyrics_clean_stopwords}
\title{Lyrics Stopwords cleaner}
\usage{
tokenize_lyrics_clean_stopwords(tokens_single)
}
\arguments{
\item{tokens_single}{A tidytext dataframe already made of unigrams}
}
\value{
A tidytext dataframe
}
\description{
Delete stopwords from unigrams tidytext dataframe
}
\examples{

}
