% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_viz.R
\name{visualize_bigrams}
\alias{visualize_bigrams}
\title{Plot graph of bigrams of a lyrics}
\usage{
visualize_bigrams(bigrams)
}
\arguments{
\item{bing_word_counts}{A tidytext dataframe without stopwords and organized in bigrams}
}
\description{
Plots graph of relationship between words, in bigrams,  of a lyrics
}
\examples{

}
