% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_viz.R
\name{viz_lda_wordcloud}
\alias{viz_lda_wordcloud}
\title{Plot comparative wordcloud for 2-topic LDA}
\usage{
viz_lda_wordcloud(ap_top_terms)
}
\arguments{
\item{ap_top_terms}{A tidytext beta dataframe already filtered}
}
\description{
Plot comparative wordcloud that shows contribute to 2-topic Latent Dirichlet Allocation
}
\examples{

}
