% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_viz.R
\name{viz_sentiment_by_artist}
\alias{viz_sentiment_by_artist}
\title{Plot sentiment of a song by artist}
\usage{
viz_sentiment_by_artist(lyric_sentiment)
}
\arguments{
\item{lyric_sentiment}{A tidytext dataframe without stopwords and joined with a sentiment database}
}
\description{
Plots the sentiment for each artist featured in the song
}
\examples{

}
