% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_viz.R
\name{viz_sentiment_word_contrib}
\alias{viz_sentiment_word_contrib}
\title{Plot Top 10 words that contribute to positive/negative sentiment of a song}
\usage{
viz_sentiment_word_contrib(bing_word_counts)
}
\arguments{
\item{bing_word_counts}{A tidytext dataframe without stopwords and joined with a sentiment database}
}
\description{
Plots the Top 10 words that contribute to positive/negative sentiment of a song
}
\examples{

}
