% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_viz.R
\name{viz_sentiment_wordcloud}
\alias{viz_sentiment_wordcloud}
\title{Plot comparative wordcloud for sentiment of words that contribute to positive/negative sentiment of a song}
\usage{
viz_sentiment_wordcloud(tokens_single_no_stop)
}
\arguments{
\item{bing_word_counts}{A tidytext dataframe without stopwords}
}
\description{
Plot comparative wordcloud that shows contribute to positive/negative sentiment of a song
}
\examples{

}
