% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_viz.R
\name{viz_top_10_words}
\alias{viz_top_10_words}
\title{Plot Top 10 words of a song}
\usage{
viz_top_10_words(tokens_single_no_stop)
}
\arguments{
\item{tokens_single_no_stop}{A tidytext dataframe without stopwords}
}
\description{
Plots the Top 10 most frequent words in a lyrics
}
\examples{

}
